Made assembler only return data related to the selected node-types
